e2e-screen-test/ログイン/requirements.txt
anthropic>=0.40.0
playwright>=1.49.0
pytest>=8.0.0
pytest-playwright>=0.6.0
pyyaml>=6.0.0
beautifulsoup4>=4.12.0
rich>=13.0.0
